+++ src/main.sh
@@ -84,6 +84,9 @@ function bashunit::main::cmd_test() {
       -R|--run-all)
         export BASHUNIT_STOP_ON_ASSERTION_FAILURE=false
         ;;
+      --preserve-env)
+        export BASHUNIT_PRESERVE_ENV=true
+        ;;
       *)
         raw_args+=("$1")
         ;;
@@ -201,6 +204,9 @@ function bashunit::main::cmd_bench() {
         ;;
       -vvv|--verbose)
         export BASHUNIT_VERBOSE=true
+        ;;
+      --preserve-env)
+        export BASHUNIT_PRESERVE_ENV=true
         ;;
       -h|--help)
         bashunit::console_header::print_bench_help